Removed variables timer and millis
Millis should not be a variable but a built-in function called by Arduino to count the milliseconds that have passed on the machine. Assigning a variable to this function will be reductant. Also the variable timer is not need because the variables timerMillis and centralStandardTime are the only times we need to function.